(+semver: fix) Reorganize code to reduce public api surface area
