+++ internal/generator/discoverTemplate.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"text/template"
 
-	"github.com/syncromatics/kafmesh/pkg/models"
+	"github.com/syncromatics/kafmesh/internal/models"
 	"github.com/syncromatics/kafmesh/pkg/runner"
 
 	"github.com/pkg/errors"
