+++ internal/generator/gen.go
@@ -8,8 +8,9 @@ import (
 	"runtime"
 	"strings"
 
+	"github.com/syncromatics/kafmesh/internal/models"
+
 	"github.com/pkg/errors"
-	"github.com/syncromatics/kafmesh/pkg/models"
 	"github.com/yargevad/filepathx"
 )
