+++ internal/generator/processorTemplate_test.go
