+++ internal/generator/serviceTemplate.go
@@ -6,8 +6,9 @@ import (
 	"path"
 	"text/template"
 
+	"github.com/syncromatics/kafmesh/internal/models"
+
 	"github.com/pkg/errors"
-	"github.com/syncromatics/kafmesh/pkg/models"
 )
 
 var (
