+++ internal/generator/sinkTemplate_test.go
