+++ internal/generator/sourceTemplate.go
@@ -4,8 +4,9 @@ import (
 	"io"
 	"text/template"
 
+	"github.com/syncromatics/kafmesh/internal/models"
+
 	"github.com/pkg/errors"
-	"github.com/syncromatics/kafmesh/pkg/models"
 )
 
 var (
