+++ internal/generator/sourceTemplate_test.go
