+++ internal/generator/topicTemplate_test.go
