+++ internal/generator/viewSinkTemplate_test.go
