+++ internal/generator/viewSourceTemplate_test.go
