+++ internal/generator/viewTemplate_test.go
