+++ pkg/runner/doc.go
@@ -1,0 +1,2 @@
+// Package runner provides the public api to kafmesh generated code.
+package runner
