+++ pkg/testing/doc.go
@@ -1,0 +1,2 @@
+// Package testing provides functions to aid in testing kafmesh services.
+package testing
